requirements.txt
fastapi
uvicorn
gunicorn
aiomysql
aiohttp
python-dotenv
editdistance
numpy